Remove tons of warnings caused by translation of C-code to C++ Remove unused functions from Getline_el.cxx; first batch. Add to history in Gl_addhistory(), not in Getlinem(). Add tab_color configuration; still need to reuse term.cxx's TTermManip object.
git-svn-id: https://root.cern.ch/svn/root/branches/dev/axel@30108 27541ba8-7e3a-0410-8455-c3a389f83636

+++ clib_editline/src/emacs.cxx
@@ -576,7 +576,7 @@ em_undo(EditLine *el, int c)
 
 		el->el_line.cursor = un->ptr;
 		size = (int) (el->el_line.cursor - el->el_line.lastchar);
-		if (size < un->isize)
+		if ((unsigned int) size < un->isize)
 			size = un->isize;
 		cp = un->ptr;
 		kp = un->buf;
